Currently:
We approximate "e ^ x" via maclaurin summation: "(x^0)/0! + (x^1)/1! + ... + (x^33)/33! + (x^34)/34!".

In order to avoid information-loss as a result of division, we first compute "(x^0)*(34!/0!) + (x^1)*(34!/1!) + ... + (x^33)*(34!/33!)", and then divide the result by "34!".

Note that we are missing the term "(x^34)*(34!/34!)" here, which means that the returned value is smaller than the desired result by "(x^34)/34!".

For "x >= 13.535535104", this yields loss of information, since "(x^34)/34! >= 1".

But this problem exists in the code from day one, so it will be handled separately (in a different branch).

-------------------------------------------------

Optimization:

The first two terms, "(x^0)*(34!/0!)" and "(x^1)*(34!/1!)", are equal to "34!" and "x*34!" respectively.

So instead of adding them to the intermediate result, which is eventually divided by "34!", we can simply add "1 + x" after dividing by "34!".

Performance improvement: substitute an assignment (=) and a multiplication (*) with an addition (+).

Accuracy improvement: since intermediate results are slightly smaller, we are able to allow for a higher maximum value of "x".

Effectively, for "0.835158750035 <= x <= 1.04290378145", we can now use "precision = 127" instead of "precision = 126".